+++ docs/release-notes.md
@@ -13,6 +13,7 @@
   * Added support for self-broadcasts through the multiplayer API. (Mods can now send messages to the current machine. That enables simple integrations between mods without needing an API, and lets mods notify a host mod without needing different code depending on whether the current player is the host or a farmhand.)
   * Added `helper.Input.GetStatus` method to get the low-level status of a button.
   * Eliminated unneeded network messages when broadcasting to a peer who can't handle the message (e.g. because they don't have SMAPI or don't have the target mod).
+  * Fixed marriage dialogue cleared when propagating dialogue changes.
 
 * For the web UI:
   * Updated the JSON validator and Content Patcher schema for `.tmx` support.